fix use 1 file _PROXY_track_cost_callback
